+++ src/index.ts
@@ -1,8 +1,10 @@
 export { transfer, completeMint } from "./transfer.js";
 export { estimateFee } from "./estimate.js";
 export { arcTestnetChain } from "./chains/arc.js";
+export { checkStellarRecipientReady } from "./chains/stellar.js";
 export { ARC_TESTNET, STELLAR_TESTNET } from "./constants.js";
 export { TransferError, SubmissionTimeoutError } from "./errors.js";
+export type { StellarRecipientStatus } from "./chains/stellar.js";
 export type {
   ChainId,
   TransferSpeed,